Move to_json function to a more generic module
